app/requirements.txt
prometheus-client
elasticsearch
pyyaml
